Move CacheAdapter and TranslationAdapter(created) into services package
Translation also needs to be used in places such as translating constant values
